rename fdcConfig file to chartConfig + minor updates to chart code
